Added watchlistmodification class to server.
Added fake functionality leveraging test data. Class will update a
watchlist item based on a username and steam id, and then update the
notification criteria for the game related to the steam id using the
passed in criteria. A dictionary is returned containing the updated
watchlist as well as whether or not the watchlist was updated.