Added gameupdater.py and watchlistupdater.py
These services are to be run from the command line when the server is offline as of right now. The watchlist updater updates the prices of all games on all watchlists while the gameupdater pulls all of the games from steam and puts them in the production game_table.json file.