Improved link steam wishlist modal dialog.
The modal dialog now only allows integers to be entered into the textfield. Also, the dialog now closes itself upon completion of the linking process. Also, on the watchlist page, when removing elements from the watchlist, after the removal the first item in the list is selected so that the remove button can be spam clicked.