Made a new listcell for watchlist games.
This listcell alerts the user to which games on the watchlist are lacking notification criteria. This was needed due to the recent feature of adding in games from a Steam wishlist. Disregarding the wishlist linking feature, this is a nice quality of life update to alert the user that they may not be getting notifications for games when they think that they should.