Completed Watchlist Sharing Feature.
On the watchlist page, the user can click the Import button. Then, a dialog appears for them to enter the username of the account whose watchlist they want to import, if the username is valid, then a new dialog is shown that displays all of the games on the entered usernames watchlist. The user can select which games they want to add to their own watchlist. If the username is invalid, then an alert is shown that tells the user that the username they entered is invalid.